+++ include/etc/tools.h
@@ -1,6 +1,7 @@
 #ifndef TOOLS_H
 #define TOOLS_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 void outb(uint16_t port, uint8_t val);
@@ -8,7 +9,7 @@ void outw(uint16_t port, uint16_t val);
 uint8_t inb(uint16_t port);
 char switch_case(char c);
 char *binrep(char *dst, uint8_t num);
-uint32_t trim(char *out, uint32_t len, const char *str);
+size_t trim(char *out, size_t len, const char *str);
 
 /*
 uint8_t inw(uint16_t port);
