+++ include/string.h
@@ -1,10 +1,11 @@
 #ifndef STRING_H
 #define STRING_H
 
+#include <stddef.h>
 #include <stdint.h>
 
-uint32_t strlen(const char *str);
+size_t strlen(const char *str);
 int strcmp(const char *str1, const char *str2);
-void *memcpy(void *dest, const void *src, uint32_t size);
+void *memcpy(void *dest, const void *src, size_t size);
 
 #endif
