+++ libc/string/strlen.c
@@ -1,7 +1,7 @@
 #include <string.h>
 
-uint32_t strlen(const char *str) {
-  uint32_t len = 0;
+size_t strlen(const char *str) {
+  size_t len = 0;
   while (str[len])
     len++;
   return len;
